added producteur & restructure
