+++ apps/templates/includes/footer.html
@@ -1,6 +1,4 @@
 <!-- /.content-wrapper -->
 <footer class="main-footer">
-  &copy; 2023
-  - coded by check-karim
-
+  &copy; 2023 - SOPAMI
 </footer>
