+++ apps/templates/includes/navigation.html
@@ -6,7 +6,7 @@
       <a class="nav-link" data-widget="pushmenu" href="#" role="button"><i class="fas fa-bars"></i></a>
     </li>
     <li class="nav-item d-none d-sm-inline-block">
-      <a href="index" class="nav-link">Home</a>
+      <a href="index" class="nav-link">Dashboard</a>
     </li>
   </ul>
 </nav>
